Add initial interfaces to installation library
