+++ src/Installer/Microsoft.Dotnet.Installation/DotnetInstallRoot.cs
@@ -1,0 +1,15 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.Dotnet.Installation;
+public record DotnetInstallRoot(
+    string? Path,
+    InstallType Type,
+    InstallArchitecture Architecture)
+{
+    //  Do we need a GUID for the ID here?
+}
