+++ src/Installer/Microsoft.Dotnet.Installation/IDotnetInstaller.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.Dotnet.Installation;
+
+public interface IDotnetInstaller
+{
+    //  Install a version of the SDK
+    //  Uninstall a version of the SDK
+}
