+++ src/Installer/Microsoft.Dotnet.Installation/IDotnetReleaseInfoProvider.cs
@@ -1,0 +1,16 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.Dotnet.Installation;
+
+public class IDotnetReleaseInfoProvider
+{
+    // Get available release channels
+    // Get latest version in a channel
+    // Get all versions in a channel
+    // Get support level for a version / channel?
+}
