+++ src/Installer/Microsoft.Dotnet.Installation/InstallArchitecture.cs
@@ -1,0 +1,15 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Microsoft.Dotnet.Installation;
+
+public enum InstallArchitecture
+{
+    x86,
+    x64,
+    arm64
+}
