--- src/Installer/dnup/IChannelVersionResolver.cs
@@ -1,12 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-using System.Collections.Generic;
-
-namespace Microsoft.DotNet.Tools.Bootstrapper
-{
-    internal interface ChannelVersionResolver
-    {
-        public DotnetInstall Resolve(DotnetInstallRequest dotnetChannelVersion);
-    }
-}
